Clean up root directory
