--- docker/nginx/project.conf
@@ -1,13 +1,0 @@
-server {
-    listen 80;
-    server_name _;
-
-    location / {
-        proxy_pass http://0.0.0.0:5000/;
-
-        # Do not change this
-        proxy_set_header Host $host;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    }
-}
