--- docker/slurk/Dockerfile_local
@@ -1,9 +1,0 @@
-FROM python:3.9
-
-RUN mkdir -p /app
-WORKDIR /app
-
-COPY requirements.txt /app
-RUN pip install --no-cache-dir -r requirements.txt
-
-ENTRYPOINT ["gunicorn", "-w", "1", "-b", ":5000", "-k", "geventwebsocket.gunicorn.workers.GeventWebSocketWorker", "run:app"]
