+++ docs/slurk_architecture.png
