+++ requirements-dev.txt
@@ -1,0 +1,4 @@
+-r requirements.txt
+docker==5.0
+pytest==6.2
+pytest-depends==1.0
