+++ requirements.txt
@@ -1,18 +1,11 @@
-docker==5.0
 Flask==2.0
 Flask-HTTPAuth==4.4
 Flask-Login==0.5
 Flask-SocketIO==5.0
 Flask-WTF==0.15.1
 Flask-Smorest==0.31.2
-gevent==21.1
-gevent-websocket
-gunicorn==20.1
 marshmallow==3.12
-pytest==6.2
-pytest-depends==1.0
 requests==2.25.1
-six
 SQLAlchemy==1.4
 Werkzeug==2.0
 WTForms==2.3
